implementing missing create account for artist, fixing description bug, redesigning card, and implementing card directory to its category page
I reinstated the create account for artists since it disappeared. The card component within the homepage was redesigned and I fixed the card overflow for business description and services. The book button directory now works and the card component is clickable and will take you to the appropriate sidebar page based off the category.