implement client card, redesign client page, styling updates for sidebar and homepage
The client page has client card imported with an accept or decline feature. Theres a filter that filters through the accepted, pending, and declined appointments. The client activity is also implemented which shows the total number of bookings, number of pending, declined, and accepted appointments. Made some more adjustments to sidebar and homepage.